Extract extractToken to module, guard string[] headers, fix PR comments
- Move extractToken to extract-token.ts so tests call the real function
  (not a copy that can silently diverge from production code)
- Guard authorization and cookie headers against string[] — Node's
  http.IncomingMessage allows duplicate headers as arrays; Authorization
  takes the first element (singular by convention), cookie joins with '; '
- Update tests to import from extract-token.ts and add two string[] cases
